requirements.txt
Flask==2.1.1
Werkzeug<2.2
Pillow>=9.5,<11.0
stegano==2.0.0
gunicorn==20.1.0
tensorflow==2.15.0
numpy>=1.23,<2.0